+++ app/services/users.py
@@ -1,12 +1,15 @@
+import json
 from http import HTTPStatus
-from typing import Optional
 
-from fastapi import Depends, HTTPException, status
+from fastapi import Depends, status
 
 from app.core.exceptions import UserAlreadyExistException, UserNotFoundException, InvalidUserCredentialsException, \
-    InvalidRolesException, UserPermissionException
+    InvalidRolesException, UserPermissionException, ActiveDepositsExistsException
 from app.core.logging import logger
 from app.core.security import verify_password, get_basic_auth_token
+from app.core.user_roles_enum import UserRoles
+from app.db.repositories.deposits import DepositsRepository, get_deposits_repository
+from app.db.repositories.products import ProductsRepository, get_products_repository
 from app.db.repositories.users import UsersRepository, get_users_repository
 from app.models.users import Users
 from app.schemas.users import UserInCreate, UserLogin, UserToken, UserInUpdate
@@ -14,9 +17,13 @@ from app.schemas.users import UserInCreate, UserLogin, UserToken, UserInUpdate
 
 class UserService:
     def __init__(
-            self, user_repo: UsersRepository = Depends(get_users_repository)
+            self, user_repo: UsersRepository = Depends(get_users_repository),
+            products_repo: ProductsRepository = Depends(get_products_repository),
+            deposits_repo: DepositsRepository = Depends(get_deposits_repository)
     ) -> None:
         self.user_repo = user_repo
+        self.products_repo = products_repo
+        self.depositions_repo = deposits_repo
 
     def login_user(self, user: UserLogin) -> UserToken:
         """
@@ -28,14 +35,18 @@ class UserService:
             token=get_basic_auth_token(username=user.username, password=user.password)
         )
 
-    def validate_roles(self, roles: list) -> list:
-        allowed_roles = {"seller", "buyer"}
-        for role in roles:
-            if role not in allowed_roles:
+    def handle_roles_update(self, user_id: int, new_roles: list, current_roles: str) -> None:
+        current_roles = json.loads(current_roles)
+        if UserRoles.SELLER_ROLE.value in current_roles and UserRoles.SELLER_ROLE.value not in new_roles:
+            if self.products_repo.are_active_products_exists_for_user(user_id):
                 raise InvalidRolesException(
-                    message=f"Invalid role: {role}. Allowed roles are: {', '.join(allowed_roles)}",
+                    message="You cannot remove seller role as there are products present under your account",
                     status_code=status.HTTP_400_BAD_REQUEST)
-        return roles
+        if UserRoles.BUYER_ROLE.value in current_roles and UserRoles.BUYER_ROLE.value not in new_roles:
+            if self.depositions_repo.get_not_utilized_deposits_by_user(user_id):
+                raise InvalidRolesException(
+                    message="You cannot remove buyer role as there are active deposits exists under your account",
+                    status_code=status.HTTP_400_BAD_REQUEST)
 
     def register_user(self, user_create: UserInCreate) -> Users:
         """
@@ -48,7 +59,6 @@ class UserService:
                 message=f"User with username: `{user_create.username}` already exists",
                 status_code=HTTPStatus.UNAUTHORIZED,
             )
-        self.validate_roles(user_create.roles)
         logger.info(f"Creating user: {user_create.username}")
         user = self.user_repo.create(obj_create=user_create)
         return user
@@ -64,8 +74,7 @@ class UserService:
                     message=f"User with username: `{user_in_update.username}` already exists",
                     status_code=HTTPStatus.UNAUTHORIZED,
                 )
-        self.validate_roles(user_in_update.roles)
-        # TODO: handle changing roles from sellers to buyers and handle products
+        self.handle_roles_update(current_user.id, user_in_update.roles, current_user.roles)
 
         user = self.user_repo.update(obj=current_user, obj_update=user_in_update)
         return user
@@ -74,7 +83,17 @@ class UserService:
         if current_user.username != username:
             raise UserPermissionException(status_code=status.HTTP_403_FORBIDDEN,
                                           message="You don't have permission to access this resource")
-        # TODO: handle deletion of users with products and transactions attached
+        current_user_roles = json.loads(current_user.roles)
+        if UserRoles.SELLER_ROLE.value in current_user_roles and self.products_repo.are_active_products_exists_for_user(
+                current_user.id):
+            return self.user_repo.disable_user(current_user)
+        if UserRoles.BUYER_ROLE.value in current_user_roles and self.depositions_repo.are_deposits_exists_for_user(
+                current_user.id):
+            if self.depositions_repo.get_not_utilized_deposits_by_user(current_user.id):
+                raise ActiveDepositsExistsException(
+                    message="There active deposits exists under your account. Kindly reset or utilize the deposits "
+                            "before proceeding for deletion of your account", status_code=status.HTTP_400_BAD_REQUEST)
+            return self.user_repo.disable_user(current_user)
         return self.user_repo.delete(current_user.id)
 
     def authenticate(self, username: str, password: str) -> Users:
@@ -83,9 +102,9 @@ class UserService:
         """
         logger.info(f"Try to authenticate user: {username}")
         user = self.user_repo.get_by_username(username=username)
-        if not user:
-            raise UserNotFoundException(
-                message=f"User with username: `{username}` not found",
+        if not user or user.disabled:
+            raise InvalidUserCredentialsException(
+                message=f"Invalid Credentials. Please try again",
                 status_code=HTTPStatus.UNAUTHORIZED,
             )
         if not verify_password(
